perf(core): Debounce concurrent cache refreshes
This is a step towards reducing the number of times that `listObjectKeys()`
is called in response to a stale cache being detected.

A loading cache as been added in front `listObjectKeys()`.

As such, loads are restricted to a single thread with all others
blocking and autocompleting after the load completes (provided the last
modified time of the object type remains constant).

This is low hanging fruit that should provide some benefit to anyone
with larger buckets (10k+ pipelines, applications, etc.) and reasonably
high request rates.