perf(core): Dont perform concurrent cache updates for a given object type
Lock object refreshes so they only occur once. If there was contention on the request, then reuse the updated cache once the contention is resolved.
The locking should not add latency since it is avoiding the work that is already happening by another thread so overall work should drop as well,
and latency for contention should be faster since the waiter doesnt have to redo work already begun.