fix(core): Fix race condition in cache invalidation
For GCS storage, we are currently deciding whether the
cache is invalid by comparing a timestamp generated in
Java (the time of last refresh) against a timestamp generated
by GCS (the time of last update). This can lead to issues with
fast read-update-read operations when the two systems don't
have perfectly synchronized clocks. Let's keep track of the last
timestamp we actually read from cloud storage, and use that to
determine whether the cache should be invalidated.