fix(core): Missing object should not trigger Hystrix fallback
There are a number of code paths where we check for the existence
of an item by calling findById and catching a NotFoundException if
thrown. Currently, when an item is not found in storage, the exception
causes us to fall back to the Hystrix fallback function. This is causing
legitimate and expected workflows to trigger the Hystrix circuit
breaker; we should not thrown an exception within Hystrix for this case.