fix(google): Fix bug in object path matching
The caching logic looks for objects ending with 'specification.json';
the logic to actually retreive the objects assumes they end with
'/specification.json', so objects such as '_specification.json' will
be listed but throw a 404 error when retreived. The listing logic
should only return objects that end in the expected
'/specification.json'.