fix(core): Warn when an object's key does not match its id
The caching logic relies on an object's key (from which its full
path in cloud storage can be derived) being equal to its key. When
this is not the case: (1) front50 will not realize it has cached the
object and will attempt to cache it every cycle, (2) attempts to
read the object by key will fail. Log an error when we're about to
cache something whose key does not match its id; in the future we
should simply skip caching such an object.