feat(app/permission): endpoint for single app permissions
This endpoint is a building block for making it possible to push new
applications into the fiat application cache when they are created.

Currently FIAT updates its application cache every 20th second and this
will help us being able to trigger updates to the cache when an
application is created.

It will follow a PR to fiat that leverages this endpoint. When both of
these are merged there will come a new front50 PR that triggers the fiat
endpoint for adding a new app so that applications will be available to
users immediately after an application is created