fix(application): Fix persistence of cloudProviders
The cloudProviders field on applications is not being persisted; it
was promoted from 'details' to its own field, but 'getPersistedProperties'
so it is not being set when initializing a new application from the JSON
that is posted to front50.  Update getPersistedProperties so it is properly
persisted and add a test case.