test(redis): Use embedded redis for unit tests
The unit tests for front50-redis depend on there being an existing
redis listening on localhost:6379.  If so, they send data to that
redis (which may not be intended to receive test data); if not, the
tests are silently skipped. Instead, set up a test embedded redis
and use that.

This also moves the unhealthy redis tests into their own class,
which autowires an unhealthy RedisConnectionFactory. Before this,
the tests were mutating the autowired bean, causing issues with
test execution order.