feat(authz): Migrate manual service users to managed service users
This migration will migrate pipelines that have manual service users to use the new permissions format. It will only run on pipelines where `roles` are not present and `runAsUser` is set to a non-managed service account.
The migration job will automatically create the new managed service users, and the new service user will get the same permissions as the manual service user that it replaces.

The migration must be manually enabled by setting the property `migrations.migrateToManagedServiceAccounts` to `true`.

*Note:* If a pipeline has multiple triggers with _different_ `runAsUser`'s set, the new managed service user will get all of the roles of the different manual service users (as you can only have one managed service user per pipeline). This can potentially remove some users ability to edit or execute affected pipelines. However, I do not think this is a very common scenario.