fix(authz): Automatic Service Account migration missing some pipelines
If a pipeline has two or more triggers, but not all are updated to automatic service accounts, the pipeline would still be skipped in the migration because `roles` would already be present. This PR fixes so that existing roles will be handled and triggers with manual service users are always migrated.