feat(core): Add endpoint to get current pipeline definition
There's currently no endpoint in front50 to get the current
pipeline definition; consumers need to call the history endpoint
with a limit of 1 and take the single returned element.

Let's make the API simpler for consumers who just want the current
definition by giving them an endpoint to return it.

Forcing consumers to call the /history endpoint has led to performance
issues, as that endpoint is often much slower than a call to get the
current definition. Some backing stores have optimized to short-circuit
the limit=1 case (s3 in #316), but others (including GCS) are still very
slow even in the limit=1 case.

We could (and potentially should) make the same optimization to other
backing stores. But I think the fundamental issue here is that we
have rare, slow history-type requests and frequent (ideally) fast
current-state requests going to the same endpoint. This PR creates
that distinction, and consumers can elect to use the new endpoint
going forward.