refactor(storage): Strengthen contract of StorageService
A StorageService should not implement any business logic and should
just store the provided item to the backing store as given.

There are a couple of places where the storage service sets a
modified timestamp (in the Azure and SQL backends); arguably this
should be done upstream so that this is independent of the backing
store but for now we'll leave those there.

Add javadoc to storeObject to explicitly strengthen the contract of
this method, which all current implementations respect. In particular,
require that the input object reflect exactly what was stored in the
persistent store upon returning from the method. Thus, implmementations
can mutatate the input object before storing it, but must ensure that
mutation is reflected on the input object.