refactor(core): Strengthen contract of StorageServiceSupport.update
Now that StorageService.storeObject explicitly guarantees that the
state of the object upon returning from the function relfects exactly
what was stored, we can strengthen the contract of
StorageServiceSupport.update.

In particular, we'll allow update to modify the supplied object (for
example to add last modified data) but will require that the state of
the input object at the end of update reflect exactly what was stored
to the persistent store.

In order to help enforce this contract, we'll make update final so that
implementations cannot override it in a way incompatible with this
contract. There were a few implementations that did override update,
but these all just did pre-update mutation/validation then called
super.update. To handle these cases, expose a preUpdate function where
implementations can specify this pre-update mutation/validation.