fix(storage): Short-circuit findById after update
Now that we have strengthened the contract of update, we can
short-circuit places where we call update to store an object,
then immediately call findById to get it back again. Given this
strengthened contract, we are guaranteed that the object in the
persistent store is identical to the object we just stored and
can avoid an extra call to the persistent store.

This is particularly important for persistent stores that don't
have strong read-after-write consistency (such as s3). In these
cases, immediately fetching the the object we just wrote may
actually return a prior version of the object (or a 404 if the
object was just created). Returning the object we know we just
wrote avoids this race condition between our call to fetch the
object and the eventual consistency of the data store.