fix(serviceAccounts): Delete managed service account when pipeline is deleted
Sometime ago, we noticed the sync operation of Fiat was very slow (it was
taking between 2 and 3 minutes) and in some cases, the Fiat instance went out
of memory. After investigations, we found out that we had about 2000 managed
service accounts not associated to any pipelines, deleting them solved our
issue with Fiat.

This change deletes the managed service account liked to the pipeline currently
being deleted.

Co-Authored-By: Greg Jones <greg.jones@gmail.com>
Co-Authored-By: Mark Vulfson <markvu@live.com>