refactor(gcs): Update retry to new format
This updates the GCS storage backend to consume the new retry
implementation that can return typed objects. This commit does the
minimal change required to get the front50 GCS implementation to work
with the new retry implementation, which is:
(1) Delete the GcsSafeRetry that extends the now final GoogleCommonSafeRetry
and just use a GoogleCommonSafeRetry directly
(2) Replace the Long values in the config with int values (as nobody is waiting
24 days between retries)
(3) Instead of storing all these retry values directly on GcsStorageService,
just create the GoogleCommonSafeRetry that uses them in the constructor
(4) Replace the catch blocks that were expecting some mix of exceptions
to now catch GoogleApiException and return a GcsProviderOperationException
wrapping it with a description of what was happening.