fix(gcs): Remove extraneous log statement
When we don't find an object, we log then throw a NotFoundException.
In general the benefit of logging then throwing an exception is
questionable; it's better to just include the details on the exception
which will either be handled or logged anyway.

In this case, there are some cases where a NotFoundException is
a reasonable outcome, and we catch it and proceed. In those cases,
the log statement is just adding noise to the logs. Let's remove it
and improve the message on the exception, so that cases that do end
up propagating or logging that exception will still get a good message
but those that choose to ignore the exception don't add noise to
the logs.