refactor(gcs): Rewrite the GcsStorageService
This rewrite uses the google-cloud-storage, which is a higher-level library than
the google-api-services-storage we were using previously. I believe the old
library is machine-generated from the protos, while the new library is manually
written by the GCP team. In addition to being simpler to use, it also comes with
some best practices baked in.

In particular for us, it includes retry logic. This means we can remove our
hand-written retry logic which both had bugs and was the source of a dependency
on clouddriver.