test(gcs): Add some implementations to FakeStorageRpc
In an upcoming commit, we'll need to also support getting a bucket;
to prepare for that let's update the FakeStorageRpc to:
* Keep track of what bucket objects are in (rather than now where it
assumes they are all in the same bucket)
* Handle creating and getting a bucket

We'll add an additional Map around the blobs data structure that
maps the bucket to a Map of blobs.  (This is a bit of a complex
data structure that in production code might be better expressed
as classes, but for this single fake in a test it's porbably fine.)

Creating a bucket initializes its blobs to an empty map; we know
if a bucket exists by whether it has an entry in the map.

I used the documentation on the interface we're mocking to decide
how to handle a missing bucket; in some cases we just return null
(when getting an object or a bucket) while in others we propagate
the 404.

This required one small change to the existing tests, which is that
we'll need to create the bucket before we use them in the tests.