fix(gcs): Fix startup error when bucketLocation is not specified
There was a regression in 1.22 where omitting bucketLocation from
one's GCS config now causes an error on startup. Prior to the
rewrite of GcsStorageService, we accepted either null or empty
string as the bucket location; now that GcsStorageService is in
kotlin and does not use a String? for the field, an error occurs
on trying to create the GcsStorageService.

The fix is just to default the string to "" in the config properties.