test(gcs): Replace nested map with classes
To simplify the nested map, define a Buckets class and a
BucketContents class, with operation names that make it more
clear what is happening.

Also, I realized that before my changes we were appending the bucket
name to the object name before putting it into the map, but now we
don't need to do that anymore because we have a separate map for
each bucket, so remove some unecessary calls to fullPath (which is
now only used in printing error messages).