fix(gcs): Fix startup error when bucketLocation is not specified (#934)
* test(gcs): Add some implementations to FakeStorageRpc

In an upcoming commit, we'll need to also support getting a bucket;
to prepare for that let's update the FakeStorageRpc to:
* Keep track of what bucket objects are in (rather than now where it
assumes they are all in the same bucket)
* Handle creating and getting a bucket

We'll add an additional Map around the blobs data structure that
maps the bucket to a Map of blobs.  (This is a bit of a complex
data structure that in production code might be better expressed
as classes, but for this single fake in a test it's porbably fine.)

Creating a bucket initializes its blobs to an empty map; we know
if a bucket exists by whether it has an entry in the map.

I used the documentation on the interface we're mocking to decide
how to handle a missing bucket; in some cases we just return null
(when getting an object or a bucket) while in others we propagate
the 404.

This required one small change to the existing tests, which is that
we'll need to create the bucket before we use them in the tests.

* test(gcs): Add test to demonstrate NPE on missing bucketLocation

This commit adds a broken test to demostrate the NPE that occurs
on startup when the bucketLocation is not specified.

* fix(gcs): Fix startup error when bucketLocation is not specified

There was a regression in 1.22 where omitting bucketLocation from
one's GCS config now causes an error on startup. Prior to the
rewrite of GcsStorageService, we accepted either null or empty
string as the bucket location; now that GcsStorageService is in
kotlin and does not use a String? for the field, an error occurs
on trying to create the GcsStorageService.

The fix is just to default the string to "" in the config properties.

* style(gcs): Use Bucket.of intead of builder

* test(gcs): Replace nested map with classes

To simplify the nested map, define a Buckets class and a
BucketContents class, with operation names that make it more
clear what is happening.

Also, I realized that before my changes we were appending the bucket
name to the object name before putting it into the map, but now we
don't need to do that anymore because we have a separate map for
each bucket, so remove some unecessary calls to fullPath (which is
now only used in printing error messages).

* style(gcs): Fix typo

* style(gcs): Fix constructor

There's no need to accept the map in a private constructor then
separately define a no-arg constructor that passes an empty map,
just make the map a val in the class.

(cherry picked from commit ba4b061812cc5a464ae00f4e07b9805b4acfde3d)