fix(sql): Ensure that `Application.createTs` is reliably returned when sql is used
For historical reasons, the original `createTs` value on the Application object itself
is not reliably stored.

When sql backends are used, the `created_at` column is the most accurate representation
of the time at which an Application was created.

This PR will override any `createTs` value that happens to exist in the Application
json blob with the value of the db column.